Deal with signedness of immed

+++ new_map/balmap.c
@@ -206,7 +206,7 @@ inline BalValue bal_int(int64_t i) {
 }
 
 inline bool bal_value_is_byte(BalValue v) {
-    return (v.immed & IMMED_TAG_INT_MASK) && v.immed <= 0x1FF;
+    return (v.immed & IMMED_TAG_INT_MASK) && (uintptr_t)v.immed <= 0x1FF;
 }
 
 inline BalValue bal_byte(uint8_t i) {
@@ -214,7 +214,7 @@ inline BalValue bal_byte(uint8_t i) {
 }
 
 inline uint8_t bal_value_to_byte_unsafe(BalValue v) {
-    assert(v.immed <= 0x1FF && (v.immed & IMMED_TAG_INT_MASK));
+    assert((uintptr_t)v.immed <= 0x1FF && (v.immed & IMMED_TAG_INT_MASK));
     return (uint8_t)(v.immed >> 1);
 }
